pyproject.toml
[tool.poetry]
name = "chuthe"
version = "0.0.0"
description = "chuthe api"
authors = ["bboyadao <anh.nguyen@pdsolution.vn>"]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
celery = { extras = ["amqp"], version = "^5.2.3" }
django-celery-results = "^2.3.0"
daphne = "^3.0.2"
django-stubs = "^1.9.0"
Markdown = "^3.3.6"
django-filter = "^22.1"
flakehell = "^0.9.0"
ipython = "^8.2.0"
whitenoise = "^6.0.0"
django-celery-beat = "^2.2.1"
pylint = "^2.14.5"
django-oauth-toolkit = "^2.0.0"
drf-spectacular = "^0.23.1"
drf-spectacular-sidecar = "^2022.4.1"
djangorestframework-stubs = "^1.7.0"
Django = "4.0.5"
coverage = "^6.4.2"
django-phonenumber-field = "^6.3.0"
phonenumbers = "^8.12.52"
dj-rest-auth = "^2.2.5"
django-cors-headers = "^3.13.0"
django-allauth = "^0.51.0"
djangorestframework-simplejwt = "^5.2.0"
firebase-admin = "^5.2.0"
fcm-django = "^1.0.12"
django-activity-stream = "^1.4.1"
JSON-log-formatter = "^0.5.1"
drf-extensions = "^0.7.1"
eventlet = "^0.33.1"
drf-nested-routers = "^0.93.4"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^5.0.0"
isort = "^5.10.1"
mypy = "^0.971"
pre-commit = "^2.20.0"
autoflake = "^1.4"
autoimport = "^1.2.2"
autopep8 = "^1.6.0"
reorder-python-imports = "^3.8.2"
yapf = "^0.32.0"
pylint = "^2.14.5"
djangorestframework-stubs = "^1.7.0"
django-extensions = "^3.2.0"
Werkzeug = "^2.1.2"

[tool.pyright]
include = ["./"]
exclude = ["**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs"
]
ignore = ["src/oldstuff"]
stubPath = "./stubs"
# venv = "$env{VIRTUAL_ENV}"

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.10"
pythonPlatform = "Linux"

executionEnvironments = [
    # { root = "src/web", pythonVersion = "3.5", pythonPlatform = "Windows", extraPaths = [ "src/service_libs" ] },
    # { root = "src/sdk", pythonVersion = "3.0", extraPaths = [ "src/backend" ] },
    # { root = "src/tests", extraPaths = ["src/tests/e2e", "src/sdk" ]},
    { root = "./" }
]

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
[tool.coverage.paths]
source = [".", ]

[tool.coverage.run]
branch = true
source = ["."]
omit = ["*/tests/*",
    "*/migrations/*",
#    "*/urls.py",
    "*/settings/*",
    "*/wsgi.py",
    "manage.py",
#    "*/apps.py",
    "*/tasks.py",
    "*/asgi.py"
]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.django-stubs]
django_settings_module = "chuthe.settings"

[tool.django-drf-stubs]
django_settings_module = "chuthe.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# list of plugins and rules for them
[tool.flakehell.plugins]
# include everything in pyflakes except F401
pyflakes = ["+*", "-F401"]
# enable only codes from S100 to S199
flake8-bandit = ["-*", "+S1??"]
# enable everything that starts from `flake8-`
"flake8-*" = ["+*"]
# explicitly disable plugin
flake8-docstrings = ["-*"]
# 80 chars limit isn't enough in 21 century
max_line_length = 120
